+++ src/core/Camera.ts
@@ -2,8 +2,14 @@
 
 import { AbstractProjection, PerspectiveProjection } from "../math/index.js";
 import { CameraStrategyFactory } from "./cameras/CameraStrategyFactory.js";
-import { CameraStrategyType } from "../enums/index.js";
-import { CameraConstraints, CameraInterfaceData, CameraStrategy } from "../interfaces/index.js";
+import { CameraEffectType, CameraStrategyType } from "../enums/index.js";
+import {
+  CameraConstraints,
+  CameraEffect,
+  CameraInterfaceData,
+  CameraStrategy,
+} from "../interfaces/index.js";
+import { CameraEffectFactory } from "./cameras/CameraEffectFactory.js";
 import { Matrix4 } from "../math/Matrix4.js";
 import { Vector3D } from "../math/Vector3D.js";
 
@@ -24,6 +30,8 @@ export class Camera implements CameraInterfaceData {
   public phi: number = 0.6;
 
   private _strategy!: CameraStrategy;
+
+  private _effects: CameraEffect[] = [];
 
   private _viewMatrix: Matrix4 = new Matrix4();
   private _viewProjMatrix: Matrix4 = new Matrix4();
@@ -63,7 +71,15 @@ export class Camera implements CameraInterfaceData {
 
   /** @inheritdoc */
   public updateViewMatrix(): void {
-    Matrix4.lookAt(this.position, this.target, this.up, this._viewMatrix);
+    const finalPos = this.position.clone();
+    const finalTarget = this.target.clone();
+
+    for (const effect of this._effects) {
+      finalPos.add(effect.offset);
+      finalTarget.add(effect.targetOffset);
+    }
+
+    Matrix4.lookAt(finalPos, finalTarget, this.up, this._viewMatrix);
     Matrix4.multiply(this.projection.getMatrix(), this._viewMatrix, this._viewProjMatrix);
   }
 
@@ -74,9 +90,9 @@ export class Camera implements CameraInterfaceData {
 
   /** @inheritdoc */
   public setStrategy(type: CameraStrategyType): void {
-    const oldConstraints = this._strategy?.constraints;
+    const oldConstraints: CameraConstraints | undefined = this._strategy?.constraints;
     this._strategy = CameraStrategyFactory.get(type);
-    if (oldConstraints) {
+    if (undefined !== oldConstraints) {
       this._strategy.constraints = oldConstraints;
     }
   }
@@ -92,7 +108,34 @@ export class Camera implements CameraInterfaceData {
   }
 
   /** @inheritdoc */
-  public update(targetPos: Vector3D, dx: number, dy: number): void {
+  public update(targetPos: Vector3D, dx: number, dy: number, deltaTime: number = 0.016): void {
     this._strategy.update(this, targetPos, dx, dy);
+
+    // Update effects
+    for (let i: number = this._effects.length - 1; 0 <= i; i--) {
+      const effect: CameraEffect = this._effects[i]!;
+      effect.update(deltaTime);
+      if (effect.isFinished) {
+        this._effects.splice(i, 1);
+      }
+    }
+  }
+
+  /**
+   * Adds a new effect to the camera.
+   * @param effect The effect to add.
+   */
+  public addEffect(effect: CameraEffect): void {
+    this._effects.push(effect);
+  }
+
+  /**
+   * Creates and adds a new effect by type.
+   * @param type The type of effect.
+   * @param intensity The intensity.
+   * @param duration The duration in seconds.
+   */
+  public applyEffect(type: CameraEffectType, intensity?: number, duration?: number): void {
+    this.addEffect(CameraEffectFactory.create(type, intensity, duration));
   }
 }